+++ Chas Advance ESP32/include/sensorDataHandler.h
@@ -4,7 +4,6 @@
 #include <Arduino.h>
 #include <ArduinoJson.h>
 #include <vector>
-#include "log.h"
 
 struct SensorData
 {
